feat: jumper 更名为 paginationMini 组件
