+++ package.json
@@ -202,6 +202,7 @@
     "dayjs": "~1.11.4",
     "hoist-non-react-statics": "~3.3.2",
     "lodash": "~4.17.15",
+    "mitt": "^3.0.0",
     "raf": "~3.4.1",
     "react-is": "^18.2.0",
     "react-popper": "~2.3.0",
