+++ src/date-picker/base/Header.tsx
@@ -3,14 +3,14 @@ import { useLocaleReceiver } from '../../locale/LocalReceiver';
 import useConfig from '../../hooks/useConfig';
 import Select from '../../select';
 import { TdDatePickerProps } from '../type';
-import Jumper, { TdJumperProps } from '../../jumper';
+import PaginationMini, { TdPaginationMiniProps } from '../../pagination';
 
 export interface DatePickerHeaderProps extends Pick<TdDatePickerProps, 'mode'> {
   year?: number;
   month?: number;
   onMonthChange?: Function;
   onYearChange?: Function;
-  onJumperClick?: TdJumperProps['onChange'];
+  onJumperClick?: TdPaginationMiniProps['onChange'];
 }
 
 const useDatePickerLocalConfig = () => {
@@ -221,7 +221,7 @@ const DatePickerHeader = (props: DatePickerHeaderProps) => {
         />
       </div>
 
-      <Jumper tips={labelMap[mode]} size="small" onChange={onJumperClick} />
+      <PaginationMini tips={labelMap[mode]} size="small" onChange={onJumperClick} />
     </div>
   );
 };
