+++ src/index.ts
@@ -56,7 +56,6 @@ export * from './range-input';
 export * from './watermark';
 export * from './image-viewer';
 export * from './space';
-export * from './jumper';
 export * from './timeline';
 export * from './image';
 export * from './rate';
