--- src/jumper/__tests__/jumper.test.tsx
@@ -1,10 +1,0 @@
-// import { render } from '@test/utils';
-// import React from 'react';
-// import Jumper from '../index';
-
-// TODO
-describe('Jumper 组件测试', () => {
-  test('dom', () => {
-    expect(true).toBe(true);
-  });
-});
