--- src/jumper/_usage/index.jsx
@@ -1,48 +1,0 @@
-/**
- *  该脚本为自动生成，如有需要请在 /script/generate-usage.js 中调整
- */
-
-// @ts-nocheck
-import React, { useState, useEffect, useMemo } from "react";
-import BaseUsage, {
-  useConfigChange,
-  usePanelChange,
-} from "@site/src/components/BaseUsage";
-import jsxToString from "react-element-to-jsx-string";
-
-import configProps from "./props.json";
-
-import { Jumper } from "tdesign-react";
-
-export default function Usage() {
-  const [configList, setConfigList] = useState(configProps);
-
-  const { changedProps, onConfigChange } = useConfigChange(configList);
-
-  const panelList = [{ label: "jumper", value: "jumper" }];
-
-  const { panel, onPanelChange } = usePanelChange(panelList);
-
-  const [renderComp, setRenderComp] = useState();
-
-  useEffect(() => {
-    setRenderComp(<Jumper {...changedProps}></Jumper>);
-  }, [changedProps]);
-
-  const jsxStr = useMemo(() => {
-    if (!renderComp) return "";
-    return jsxToString(renderComp);
-  }, [renderComp]);
-
-  return (
-    <BaseUsage
-      code={jsxStr}
-      panelList={panelList}
-      configList={configList}
-      onPanelChange={onPanelChange}
-      onConfigChange={onConfigChange}
-    >
-      {renderComp}
-    </BaseUsage>
-  );
-}
