--- src/jumper/_usage/props.json
@@ -1,63 +1,0 @@
-[
-  {
-    "name": "disabled",
-    "type": "Boolean",
-    "defaultValue": false,
-    "options": []
-  },
-  {
-    "name": "layout",
-    "type": "enum",
-    "defaultValue": "horizontal",
-    "options": [
-      {
-        "label": "horizontal",
-        "value": "horizontal"
-      },
-      {
-        "label": "vertical",
-        "value": "vertical"
-      }
-    ]
-  },
-  {
-    "name": "showCurrent",
-    "type": "Boolean",
-    "defaultValue": true,
-    "options": []
-  },
-  {
-    "name": "size",
-    "type": "enum",
-    "defaultValue": "medium",
-    "options": [
-      {
-        "label": "small",
-        "value": "small"
-      },
-      {
-        "label": "medium",
-        "value": "medium"
-      },
-      {
-        "label": "large",
-        "value": "large"
-      }
-    ]
-  },
-  {
-    "name": "variant",
-    "type": "enum",
-    "defaultValue": "text",
-    "options": [
-      {
-        "label": "text",
-        "value": "text"
-      },
-      {
-        "label": "outline",
-        "value": "outline"
-      }
-    ]
-  }
-]
