--- src/jumper/defaultProps.ts
@@ -1,12 +1,0 @@
-/**
- * 该文件为脚本自动生成文件，请勿随意修改。如需修改请联系 PMC
- * */
-
-import { TdJumperProps } from './type';
-
-export const jumperDefaultProps: TdJumperProps = {
-  layout: 'horizontal',
-  showCurrent: true,
-  size: 'medium',
-  variant: 'text',
-};
