--- src/jumper/index.ts
@@ -1,9 +1,0 @@
-import _Jumper from './Jumper';
-
-import './style/index.js';
-
-export type { JumperProps } from './Jumper';
-export * from './type';
-
-export const Jumper = _Jumper;
-export default Jumper;
