--- src/jumper/style/css.js
@@ -1,1 +1,0 @@
-import './index.css';
