--- src/jumper/style/index.js
@@ -1,1 +1,0 @@
-import '../../_common/style/web/components/jumper/_index.less';
